Update main.py
Some improvements with amounts to bridge. 
Now the script will check each time whether the balance of USDC on the wallet corresponds to the amount you want to exchange. And if the amount you want to exchange will be more than the amount you have in your wallet, the script will change the amount that you put in main.py to the amount of USDC in your wallet.

The amount will change only if you have less USDC in your wallet than you specified for the transaction. And the amount of USDC will gradually become less, because the bridge itself eats up a small part when transferring from one network to another.